SheetUpdater/requirements.txt
psycopg2-binary
schedule
gspread
oauth2client
schedule